Create dummy data for testing the completions functions

Creates a prepared dataframe to test scripts related to IPEDS Completions reporting. Produces either a student/degree dataframe or a dataframe of cips previously reported but not in the current student data, depending on the argument you select

create_dummy_data_com(df_type = "student")

Arguments

df_type

a string: "student" to get the main df needed, "cip" to get extracips

Value

a dataframe ready for the rest of the comp scripts

Details

remember: want to save this data out into the package so it's available

Note

The final dataset has 100 students with 190 majors. 60 students have 1 major for 1 degree 8 students have 2 majors for 1 degree 17 students have 3 majors (on one degree or as dual-degree) 12 students have 4 majors (as dual-degrees; some have one as an MBA) 3 students have 5 majors (dual-degree and 4 majors and an MBA) Note: 1 student has a faulty birthdate; this will show the warning "1 failed to parse" I use the same seed for each grouping, to ensure that the same students keep getting more things added One program/level combination is flagged as distance education With the seed and probs, no associates degrees end up in the data (don't worry!) To fully process completions, we will need to include an example of a CIP code that is a possible major but has no completers and a CIP code in an award level that is possible but has no completers This is the second piece of dummy df produced

Examples

set.seed(1892)
create_dummy_data_com()
#> Warning:  1 failed to parse.
#>     Unitid StudentId RaceEthnicity Sex DegreeLevel MajorNumber MajorCip
#> 1   999999       100             4   2           5           1   9.0100
#> 2   999999       101             8   2           7           1   9.0401
#> 3   999999       102             7   2           5           1   9.0999
#> 4   999999       103             6   1           2           1   9.9999
#> 5   999999       104             7   2           5           1   9.0999
#> 6   999999       105             6   2           5           1   9.1001
#> 7   999999       106             4   1           5           1   9.0100
#> 8   999999       107             7   2           7           1   9.0901
#> 9   999999       108             3   1           5           1   9.0100
#> 10  999999       109             2   2           4           1   9.9999
#> 11  999999       110             5   2           5           1   9.1001
#> 12  999999       111             4   2           7           1   9.9999
#> 13  999999       112             2   1          17           1   9.0701
#> 14  999999       113             1   1          18           1   9.0100
#> 15  999999       114             5   2           6           1   9.9999
#> 16  999999       115             6   2           5           1   9.9999
#> 17  999999       116             9   2           5           1   9.0901
#> 18  999999       117             5   2           5           1   9.0999
#> 19  999999       118             1   2           5           1   9.0100
#> 20  999999       119             3   2          17           1   9.0999
#> 21  999999       120             5   2           7           1   9.0901
#> 22  999999       121             6   2           7           1   9.1001
#> 23  999999       122             8   1           5           1   9.0901
#> 24  999999       123             1   2           5           1   9.1001
#> 25  999999       124             7   2           5           1   9.0701
#> 26  999999       125             7   1          17           1   9.0701
#> 27  999999       126             8   2          17           1   9.0100
#> 28  999999       127             6   1           5           1   9.0100
#> 29  999999       128             5   2           5           1   9.0100
#> 30  999999       129             9   1           7           1   9.0901
#> 31  999999       130             1   1           5           1   9.0901
#> 32  999999       131             2   1           2           1   9.0901
#> 33  999999       132             5   2           7           1   9.9999
#> 34  999999       133             7   1           7           1   9.0100
#> 35  999999       134             9   2           7           1   9.1001
#> 36  999999       135             6   1           7           1   9.0701
#> 37  999999       136             2   2           7           1   9.9999
#> 38  999999       137             2   2           7           1   9.0999
#> 39  999999       138             7   1           7           1   9.0701
#> 40  999999       139             1   1           7           1   9.0100
#> 41  999999       140             5   1          18           1   9.0701
#> 42  999999       141             6   1           5           1   9.0401
#> 43  999999       142             6   2           5           1   9.0901
#> 44  999999       143             6   2           5           1   9.0999
#> 45  999999       144             4   2           5           1   9.9999
#> 46  999999       145             9   2           5           1   9.0401
#> 47  999999       146             4   2           5           1   9.0100
#> 48  999999       147             6   2           7           1   9.0100
#> 49  999999       148             1   1           5           1   9.0100
#> 50  999999       149             4   1           5           1   9.9999
#> 51  999999       150             9   1          17           1   9.0999
#> 52  999999       151             8   1           5           1   9.9999
#> 53  999999       152             1   2           5           1   9.1001
#> 54  999999       153             2   1           5           1   9.9999
#> 55  999999       154             5   1          18           1   9.9999
#> 56  999999       155             1   2          17           1   9.1001
#> 57  999999       156             2   1           7           1   9.0999
#> 58  999999       157             1   2           5           1   9.0901
#> 59  999999       158             6   1          18           1   9.0401
#> 60  999999       159             4   1          18           1   9.0100
#> 61  999999       160             3   2           5           1   9.9999
#> 62  999999       161             2   2           7           1   9.0100
#> 63  999999       162             1   1           5           1   9.0100
#> 64  999999       163             7   1           5           1   9.1001
#> 65  999999       164             5   1          18           1   9.0999
#> 66  999999       165             5   2           5           1   9.0401
#> 67  999999       166             5   2           5           1   9.0401
#> 68  999999       167             5   2           7           1   9.9999
#> 69  999999       168             4   1           7           1   9.0701
#> 70  999999       169             1   1          19           1   9.0100
#> 71  999999       170             7   1           5           1   9.0100
#> 72  999999       171             6   1           7           1   9.0701
#> 73  999999       172             3   1           7           1   9.0401
#> 74  999999       173             2   1          19           1   9.0701
#> 75  999999       174             8   2           7           1   9.0701
#> 76  999999       175             4   2           7           1   9.0701
#> 77  999999       176             3   2           5           1   9.0901
#> 78  999999       177             5   2           7           1   9.1001
#> 79  999999       178             2   2          19           1   9.0999
#> 80  999999       179             6   2          17           1   9.0999
#> 81  999999       180             2   1           7           1   9.0999
#> 82  999999       181             5   2          18           1   9.0701
#> 83  999999       182             1   1           7           1   9.0701
#> 84  999999       183             2   1          17           1   9.0100
#> 85  999999       184             5   1           4           1   9.0901
#> 86  999999       185             7   2           5           1   9.0999
#> 87  999999       186             1   2           5           1   9.0701
#> 88  999999       187             9   2           7           1   9.0901
#> 89  999999       188             3   2           5           1   9.0901
#> 90  999999       189             7   2           5           1   9.0701
#> 91  999999       190             2   2          17           1   9.0701
#> 92  999999       191             5   2           5           1   9.0901
#> 93  999999       192             4   2           7           1   9.0100
#> 94  999999       193             5   1           5           1   9.0901
#> 95  999999       194             6   2           7           1   9.0701
#> 96  999999       195             6   2           5           1   9.9999
#> 97  999999       196             5   2           2           1   9.0701
#> 98  999999       197             9   1           5           1   9.0100
#> 99  999999       198             6   1           5           1   9.0701
#> 100 999999       199             3   2           5           1   9.0701
#> 101 999999       106             4   1           5           2  31.0301
#> 102 999999       108             3   1           5           2  31.0504
#> 103 999999       110             5   2           5           2  31.0399
#> 104 999999       112             2   1          17           2  31.0504
#> 105 999999       116             9   2           5           2  31.0101
#> 106 999999       117             5   2           5           2  31.0504
#> 107 999999       119             3   2          17           2  31.0301
#> 108 999999       124             7   2           5           2  31.0504
#> 109 999999       129             9   1           7           2  31.0101
#> 110 999999       132             5   2           7           2  31.0101
#> 111 999999       133             7   1           7           2  31.0504
#> 112 999999       134             9   2           7           2  31.0504
#> 113 999999       135             6   1           7           2  31.0399
#> 114 999999       136             2   2           7           2  31.0101
#> 115 999999       140             5   1          18           2  31.0101
#> 116 999999       146             4   2           5           2  31.0301
#> 117 999999       147             6   2           7           2  31.0101
#> 118 999999       149             4   1           5           2  31.0101
#> 119 999999       151             8   1           5           2  31.0101
#> 120 999999       153             2   1           5           2  31.0101
#> 121 999999       156             2   1           7           2  31.0301
#> 122 999999       159             4   1          18           2  31.0301
#> 123 999999       160             3   2           5           2  31.0399
#> 124 999999       164             5   1          18           2  31.0504
#> 125 999999       167             5   2           7           2  31.0399
#> 126 999999       170             7   1           5           2  31.0399
#> 127 999999       176             3   2           5           2  31.0301
#> 128 999999       177             5   2           7           2  31.0399
#> 129 999999       178             2   2          19           2  31.0504
#> 130 999999       179             6   2          17           2  31.0101
#> 131 999999       182             1   1           7           2  31.0399
#> 132 999999       185             7   2           5           2  31.0101
#> 133 999999       186             1   2           5           2  31.0504
#> 134 999999       187             9   2           7           2  31.0101
#> 135 999999       189             7   2           5           2  31.0101
#> 136 999999       191             5   2           5           2  31.0101
#> 137 999999       193             5   1           5           2  31.0399
#> 138 999999       100             4   2           5           1  14.0702
#> 139 999999       105             6   2           5           1  14.0101
#> 140 999999       110             5   2           5           1  14.0101
#> 141 999999       118             1   2           5           1  14.0799
#> 142 999999       121             6   2           7           1  14.0799
#> 143 999999       124             7   2           5           1  14.0799
#> 144 999999       129             9   1           7           1  14.0801
#> 145 999999       138             7   1           7           1  14.0702
#> 146 999999       139             1   1           7           1  14.0799
#> 147 999999       140             5   1          18           1  14.0101
#> 148 999999       142             6   2           5           1  14.0702
#> 149 999999       145             9   2           5           1  14.0702
#> 150 999999       149             4   1           5           1  14.0101
#> 151 999999       150             9   1          17           1  14.0801
#> 152 999999       151             8   1           5           1  14.0801
#> 153 999999       156             2   1           7           1  14.0799
#> 154 999999       157             1   2           5           1  14.0801
#> 155 999999       158             6   1          18           1  14.0101
#> 156 999999       159             4   1          18           1  14.0101
#> 157 999999       160             3   2           5           1  14.0799
#> 158 999999       162             1   1           5           1  14.0801
#> 159 999999       165             5   2           5           1  14.0101
#> 160 999999       168             4   1           7           1  14.0101
#> 161 999999       180             2   1           7           1  14.0801
#> 162 999999       188             3   2           5           1  14.0101
#> 163 999999       193             5   1           5           1  14.0801
#> 164 999999       195             6   2           5           1  14.0801
#> 165 999999       199             3   2           5           1  14.0702
#> 166 999999       106             4   1           5           2  16.0199
#> 167 999999       117             5   2           5           2  16.0199
#> 168 999999       119             3   2          17           2  16.0104
#> 169 999999       129             9   1           7           2  16.0104
#> 170 999999       135             6   1           7           2  16.0102
#> 171 999999       140             5   1          18           2  16.0199
#> 172 999999       146             4   2           5           2  16.0105
#> 173 999999       149             4   1           5           2  16.0199
#> 174 999999       156             2   1           7           2  16.0105
#> 175 999999       164             5   1          18           2  16.0105
#> 176 999999       177             5   2           7           2  16.0105
#> 177 999999       185             7   2           5           2  16.0105
#> 178 999999       186             1   2           5           2  16.0199
#> 179 999999       187             9   2           7           2  16.0102
#> 180 999999       189             7   2           5           2  16.0104
#> 181 999999       191             5   2           5           2  16.0104
#> 182 999999       106             4   1           7           1  52.0201
#> 183 999999       131             2   1           7           1  52.0201
#> 184 999999       162             1   1           7           1  52.0201
#> 185 999999       166             5   2           7           1  52.0201
#> 186 999999       173             2   1           7           1  52.0201
#> 187 999999       177             5   2           7           1  52.0201
#> 188 999999       180             2   1           7           1  52.0201
#> 189 999999       191             5   2           7           1  52.0201
#> 190 999999       193             5   1           7           1  52.0201
#> 191 999999       198             6   1           7           1  52.0201
#>     DistanceEd DistanceEd31 DistanceEd32  Birthdate      Age
#> 1            2           NA           NA       <NA>       NA
#> 2            2           NA           NA 1999-01-04 19.36208
#> 3            2           NA           NA 1999-09-26 18.63655
#> 4            2           NA           NA 1999-07-23 18.81451
#> 5            2           NA           NA 1977-05-08 41.02122
#> 6            2           NA           NA 1999-08-13 18.75702
#> 7            2           NA           NA 1999-07-22 18.81725
#> 8            2           NA           NA 1988-01-16 30.32991
#> 9            2           NA           NA 1977-08-16 40.74743
#> 10           2           NA           NA 1988-01-10 30.34634
#> 11           2           NA           NA 1999-03-21 19.15400
#> 12           2           NA           NA 1999-01-20 19.31828
#> 13           2           NA           NA 1999-08-07 18.77344
#> 14           2           NA           NA 1999-06-01 18.95688
#> 15           2           NA           NA 1999-03-06 19.19507
#> 16           2           NA           NA 1988-11-23 29.47570
#> 17           2           NA           NA 1999-05-16 19.00068
#> 18           2           NA           NA 1999-01-25 19.30459
#> 19           2           NA           NA 1999-02-04 19.27721
#> 20           2           NA           NA 2002-01-30 16.29021
#> 21           2           NA           NA 1999-11-17 18.49418
#> 22           2           NA           NA 1999-09-30 18.62560
#> 23           2           NA           NA 1999-06-14 18.92129
#> 24           2           NA           NA 1999-04-10 19.09925
#> 25           2           NA           NA 1999-05-21 18.98700
#> 26           2           NA           NA 1999-02-22 19.22793
#> 27           2           NA           NA 1999-09-01 18.70500
#> 28           2           NA           NA 1999-07-04 18.86653
#> 29           2           NA           NA 1999-11-05 18.52704
#> 30           2           NA           NA 1999-05-05 19.03080
#> 31           2           NA           NA 1999-07-06 18.86105
#> 32           2           NA           NA 1999-10-03 18.61739
#> 33           2           NA           NA 1977-02-14 41.24846
#> 34           2           NA           NA 1999-03-17 19.16496
#> 35           2           NA           NA 1999-07-14 18.83915
#> 36           2           NA           NA 1999-04-14 19.08830
#> 37           2           NA           NA 1999-06-21 18.90212
#> 38           2           NA           NA 1988-03-31 30.12457
#> 39           2           NA           NA 1999-10-14 18.58727
#> 40           2           NA           NA 1999-01-29 19.29363
#> 41           2           NA           NA 1999-11-19 18.48871
#> 42           2           NA           NA 1999-02-20 19.23340
#> 43           2           NA           NA 1999-04-17 19.08008
#> 44           2           NA           NA 1999-04-24 19.06092
#> 45           2           NA           NA 1999-07-20 18.82272
#> 46           2           NA           NA 1999-11-07 18.52156
#> 47           2           NA           NA 1999-04-06 19.11020
#> 48           2           NA           NA 1999-12-08 18.43669
#> 49           2           NA           NA 2002-03-12 16.17796
#> 50           2           NA           NA 1999-06-12 18.92676
#> 51           2           NA           NA 1999-09-25 18.63929
#> 52           2           NA           NA 1977-01-25 41.30322
#> 53           2           NA           NA 1999-05-20 18.98973
#> 54           2           NA           NA 1999-08-07 18.77344
#> 55           2           NA           NA 1988-10-09 29.59890
#> 56           2           NA           NA 1999-01-26 19.30185
#> 57           2           NA           NA 1999-06-07 18.94045
#> 58           2           NA           NA 1999-07-29 18.79808
#> 59           2           NA           NA 1999-12-04 18.44764
#> 60           2           NA           NA 1999-02-11 19.25804
#> 61           2           NA           NA 1999-06-11 18.92950
#> 62           2           NA           NA 1999-08-29 18.71321
#> 63           2           NA           NA 1988-12-09 29.43190
#> 64           2           NA           NA 1999-10-10 18.59822
#> 65           2           NA           NA 1999-06-14 18.92129
#> 66           2           NA           NA 1977-12-05 40.44353
#> 67           2           NA           NA 1999-02-04 19.27721
#> 68           2           NA           NA 1999-01-29 19.29363
#> 69           2           NA           NA 1999-03-12 19.17864
#> 70           2           NA           NA 1999-06-07 18.94045
#> 71           2           NA           NA 1999-03-16 19.16769
#> 72           2           NA           NA 1999-05-31 18.95962
#> 73           2           NA           NA 1999-03-04 19.20055
#> 74           2           NA           NA 1999-10-04 18.61465
#> 75           2           NA           NA 1999-11-07 18.52156
#> 76           2           NA           NA 1988-08-28 29.71389
#> 77           2           NA           NA 1999-06-26 18.88843
#> 78           2           NA           NA 1999-08-15 18.75154
#> 79           2           NA           NA 1999-02-08 19.26626
#> 80           2           NA           NA 1999-01-17 19.32649
#> 81           2           NA           NA 1977-02-04 41.27584
#> 82           2           NA           NA 1999-05-22 18.98426
#> 83           2           NA           NA 1999-01-13 19.33744
#> 84           2           NA           NA 1999-01-09 19.34839
#> 85           2           NA           NA 1999-08-02 18.78713
#> 86           2           NA           NA 1988-08-31 29.70568
#> 87           2           NA           NA 1999-03-01 19.20876
#> 88           2           NA           NA 1999-12-02 18.45311
#> 89           2           NA           NA 1999-12-15 18.41752
#> 90           2           NA           NA 1999-01-12 19.34018
#> 91           2           NA           NA 1999-05-17 18.99795
#> 92           2           NA           NA 1988-09-22 29.64545
#> 93           2           NA           NA 1999-05-07 19.02533
#> 94           2           NA           NA 1999-03-22 19.15127
#> 95           2           NA           NA 1999-12-23 18.39562
#> 96           2           NA           NA 1999-09-01 18.70500
#> 97           2           NA           NA 1999-12-01 18.45585
#> 98           2           NA           NA 1999-10-28 18.54894
#> 99           2           NA           NA 1999-04-13 19.09103
#> 100          2           NA           NA 1999-07-15 18.83641
#> 101          2           NA           NA 1999-07-22 18.81725
#> 102          2           NA           NA 1977-08-16 40.74743
#> 103          2           NA           NA 1999-03-21 19.15400
#> 104          2           NA           NA 1999-08-07 18.77344
#> 105          2           NA           NA 1999-05-16 19.00068
#> 106          2           NA           NA 1999-01-25 19.30459
#> 107          2           NA           NA 2002-01-30 16.29021
#> 108          2           NA           NA 1999-05-21 18.98700
#> 109          2           NA           NA 1999-05-05 19.03080
#> 110          2           NA           NA 1977-02-14 41.24846
#> 111          2           NA           NA 1999-03-17 19.16496
#> 112          2           NA           NA 1999-07-14 18.83915
#> 113          2           NA           NA 1999-04-14 19.08830
#> 114          2           NA           NA 1999-06-21 18.90212
#> 115          2           NA           NA 1999-11-19 18.48871
#> 116          2           NA           NA 1999-04-06 19.11020
#> 117          2           NA           NA 1999-12-08 18.43669
#> 118          2           NA           NA 1999-06-12 18.92676
#> 119          2           NA           NA 1977-01-25 41.30322
#> 120          2           NA           NA 1999-08-07 18.77344
#> 121          2           NA           NA 1999-06-07 18.94045
#> 122          2           NA           NA 1999-02-11 19.25804
#> 123          2           NA           NA 1999-06-11 18.92950
#> 124          2           NA           NA 1999-06-14 18.92129
#> 125          2           NA           NA 1999-01-29 19.29363
#> 126          2           NA           NA 1999-03-16 19.16769
#> 127          2           NA           NA 1999-06-26 18.88843
#> 128          2           NA           NA 1999-08-15 18.75154
#> 129          2           NA           NA 1999-02-08 19.26626
#> 130          2           NA           NA 1999-01-17 19.32649
#> 131          2           NA           NA 1999-01-13 19.33744
#> 132          2           NA           NA 1988-08-31 29.70568
#> 133          2           NA           NA 1999-03-01 19.20876
#> 134          2           NA           NA 1999-12-02 18.45311
#> 135          2           NA           NA 1999-01-12 19.34018
#> 136          2           NA           NA 1988-09-22 29.64545
#> 137          2           NA           NA 1999-03-22 19.15127
#> 138          2           NA           NA       <NA>       NA
#> 139          2           NA           NA 1999-08-13 18.75702
#> 140          2           NA           NA 1999-03-21 19.15400
#> 141          2           NA           NA 1999-02-04 19.27721
#> 142          2           NA           NA 1999-09-30 18.62560
#> 143          2           NA           NA 1999-05-21 18.98700
#> 144          2           NA           NA 1999-05-05 19.03080
#> 145          2           NA           NA 1999-10-14 18.58727
#> 146          2           NA           NA 1999-01-29 19.29363
#> 147          2           NA           NA 1999-11-19 18.48871
#> 148          2           NA           NA 1999-04-17 19.08008
#> 149          2           NA           NA 1999-11-07 18.52156
#> 150          2           NA           NA 1999-06-12 18.92676
#> 151          2           NA           NA 1999-09-25 18.63929
#> 152          2           NA           NA 1977-01-25 41.30322
#> 153          2           NA           NA 1999-06-07 18.94045
#> 154          2           NA           NA 1999-07-29 18.79808
#> 155          2           NA           NA 1999-12-04 18.44764
#> 156          2           NA           NA 1999-02-11 19.25804
#> 157          2           NA           NA 1999-06-11 18.92950
#> 158          2           NA           NA 1988-12-09 29.43190
#> 159          2           NA           NA 1977-12-05 40.44353
#> 160          2           NA           NA 1999-03-12 19.17864
#> 161          2           NA           NA 1977-02-04 41.27584
#> 162          2           NA           NA 1999-12-15 18.41752
#> 163          2           NA           NA 1999-03-22 19.15127
#> 164          2           NA           NA 1999-09-01 18.70500
#> 165          2           NA           NA 1999-07-15 18.83641
#> 166          2           NA           NA 1999-07-22 18.81725
#> 167          2           NA           NA 1999-01-25 19.30459
#> 168          2           NA           NA 2002-01-30 16.29021
#> 169          2           NA           NA 1999-05-05 19.03080
#> 170          2           NA           NA 1999-04-14 19.08830
#> 171          2           NA           NA 1999-11-19 18.48871
#> 172          2           NA           NA 1999-04-06 19.11020
#> 173          2           NA           NA 1999-06-12 18.92676
#> 174          2           NA           NA 1999-06-07 18.94045
#> 175          2           NA           NA 1999-06-14 18.92129
#> 176          2           NA           NA 1999-08-15 18.75154
#> 177          2           NA           NA 1988-08-31 29.70568
#> 178          2           NA           NA 1999-03-01 19.20876
#> 179          2           NA           NA 1999-12-02 18.45311
#> 180          2           NA           NA 1999-01-12 19.34018
#> 181          2           NA           NA 1988-09-22 29.64545
#> 182          3            1            0 1999-07-22 18.81725
#> 183          3            1            0 1999-10-03 18.61739
#> 184          3            1            0 1988-12-09 29.43190
#> 185          3            1            0 1999-02-04 19.27721
#> 186          3            1            0 1999-10-04 18.61465
#> 187          3            1            0 1999-08-15 18.75154
#> 188          3            1            0 1977-02-04 41.27584
#> 189          3            1            0 1988-09-22 29.64545
#> 190          3            1            0 1999-03-22 19.15127
#> 191          3            1            0 1999-04-13 19.09103
create_dummy_data_com(df_type = "cip")
#> Warning:  3 failed to parse.
#> Joining, by = c("MajorCip", "DegreeLevel", "DistanceEd", "DistanceEd31",
#> "DistanceEd32")
#>    Unitid MajorNumber MajorCip DegreeLevel DistanceEd DistanceEd31 DistanceEd32
#> 11 999999           1  45.1001          1b          1           NA           NA
#> 21 999999           1  45.1001           5          1           NA           NA
#> 14 999999           1  22.0206          1a          3            0            1
#>    RaceEthnicity Sex Count
#> 11             1   1     0
#> 21             1   1     0
#> 14             1   1     0