Output files from the make functions using the package dummy data (for testing)

Contains a list of output files from each "make" script which can be called by testthat tests to ensure refactoring doesn't change output inappropriately

part_outputs

Format

A list of dataframes: one dataframe per make script

Details

General list-creation code is commented out in the test-part-outputs.R script; uncomment and run/add as needed if a df needs a legitimate update (e.g. IPEDS rules change, dummy data includes new use case)

Examples

part_outputs$com_partD
#> Error in eval(expr, envir, enclos): object 'part_outputs' not found