Shortcut function with all steps to provide a Fall Enrollment report

Shortcut function with all steps to provide a Fall Enrollment report

Usage

produce_ef1_report(
  students,
  retention,
  part = "ALL",
  include_optional = FALSE,
  format = "uploadable",
  ugender = TRUE,
  ggender = TRUE
)

Arguments

students

A dataframe set up according to the readme with student data

retention

A dataframe set up according to the readme with retention data

part

A string with what part of the report you want to produce: 'all', 'A', etc.

include_optional

A boolean flag for whether optional parts should be included

format

A string ("uploadable" will produce a properly formatted upload file. "readable" will produce a csv of the upload file (only works for one part at a time). "both" will provide both options, but only works with one part at a time.

ugender

A boolean: TRUE means you are collecting and able to report "another gender" for undergraduate completers, even if you have no (or few) such students. Set as FALSE if necessary

ggender

A boolean: TRUE means you are collecting and able to report "another gender" for graduate completers, even if you have no (or few) such students. Set as FALSE if necessary

Value

A txt or csv file at the path of your choice

Examples

# \donttest{
#entire report
produce_ef1_report(ef1_students, ef1_retention)
#> Uploadable results available at /private/var/folders/m4/5dz5h26x329cqq4fx333f8gm0000gn/T/RtmppKCnCQ/FallEnrollment_AllParts_2024-08-30.txt

#entire report with optional sections
produce_ef1_report(ef1_students, ef1_retention, include_optional = TRUE)
#> Uploadable results available at /private/var/folders/m4/5dz5h26x329cqq4fx333f8gm0000gn/T/RtmppKCnCQ/FallEnrollment_AllParts_2024-08-30.txt

#one part as csv instead of key-value
produce_ef1_report(ef1_students, part = 'D', format = 'readable')
#> Readable results available at  /private/var/folders/m4/5dz5h26x329cqq4fx333f8gm0000gn/T/RtmppKCnCQ/FallEnrollment_PartD_Readable_2024-08-30.csv
# }